store: replace Modified+Load with ReloadIfChanged
simplify the pattern if store.Modified() => store.Load() with a single
function call store.ReloadIfChanged().

In addition to the code simplification, it solves a race condition
that when only a RLock() is hold and that multiple goroutines could
run the sequence store.Modified() and store.Load().  In this case only
one goroutine would see store.Modified( and run store.Load().

Other goroutines could access the old data while the goroutine that
saw the change modifies the store data.

Closes: https://github.com/containers/storage/issues/880

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>