containers.conf: don't set default logging driver
After experimenting with defaulting to journald as the default logging
driver, we have reverted to k8s-file in the built-in defaults in memory
but left journald the default in the containers.conf.

To prevent surprised in downstream packaging which may pick up the
containers.conf file, let's revert it as well.

Signed-off-by: Valentin Rothberg <rothberg@redhat.com>