storage: set image names last
Set the names on an image last, after we've stored its config blob,
manifests, and signatures.  This will keep callers that look up an image
and then immediately try to inspect it from getting tripped up if they
hit the window between when we set the names and save the data that's
used when we inspect an image.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>