Avoid failure when unmount an unmounted mountpoint
Check if the mountpoint is mounted when unmount it to avoid failure.
If user manually run the umount command before it, the function
UnmountDevice returns an error, Although this error dose not cause the
container deletion process fail for the reason that the return value of
UnmountDevice function is not processed. However, the ERROR logs in the
log system are misleading

Signed-off-by: gaohuatao <gaohuatao@huawei.com>