store: fix graphLock reload
use s.graphLock.Modified() instead of s.graphLock.TouchedSince().

TouchedSince() seems to fail in some cases when comparing the inode
mtime with the current time.

Avoid such kind of problems in a critical code path as store.mount(),
as we must be sure there is already a driver home mount present,
otherwise it the next process may cover the container mount with the
home mount.

Closes: https://github.com/containers/podman/issues/10454

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>