copy: re-enable early commits
The early committing of blobs has been temporarily disabled in #1205.
There was a regression in deciding whether an empty layer may be
committed to the storage or not [1].

Initially, the early committing considerred a layer to empty (and hence
not commit it to storage) when it matched the shasum of an empty GZIP
layer.  As it turned out, that is not a reliable indicator since some
images may still include these layers in the diff IDs [2].

Re-enable the early committing of blobs by passing down the emptiness
information down by querrying the image's manifest.

Signed-off-by: Valentin Rothberg <rothberg@redhat.com>

[1] https://github.com/containers/image/pull/1205#issuecomment-822459523
[2] https://github.com/containers/image/pull/1205#issuecomment-822458565

Signed-off-by: Valentin Rothberg <rothberg@redhat.com>