Reload layer storage if layers.json got externally modified
This helps long running processes like CRI-O to determine changes to the
local storage, while handling corrupted images automatically.

The corresponding fix in Podman [0] handles corrupt layers by reloading
the image. This does not work for CRI-O, because it will not reload the
storage on subsequent calls of pullImage if no storage modification has
been done.

[0]: https://github.com/containers/podman/pull/10637/commits/b4bd886fcc01c99b1c8aa5800710b1c31bae8ae5

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>