storage: PutBlobWithOptions: fix EmptyLayer logic
As it turned out in containers/podman/pull/10700,
containers/image@9b0e8f7 introduced a bug in the put-blob logic where
the empty-layer info was used for all blobs that the worker will commit
to storage.

Fix the bug by correctly storing the empty-layer information for each
blob separately and use that inside the worker.

Signed-off-by: Valentin Rothberg <rothberg@redhat.com>