Use orderedSet in copier.determineListConversion
The extra hash table it maintains is a bit more expensive,
but it allows us to directly express our preferences, and
it is more similar to determineManifestConversion
(if we ever wanted to add a prioritization for
manifest list formats).

Also use the two-argument for...range loops to avoid
repetition of the long array names.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>