pull: fallthrough for registry parsing errors
Pull is a bit of an "one size fits all" API to keep complexity away from
callers and hide everything behind the interface.

Commit 369aaa4178e7 recently altered the error reporting to when pulling
fromt the `docker-daemon` transport which in turn caused a regression in
Buildah CI when pulling `docker:latest`.  Such an input would cause a
parsing error in the `docker:`.

Fix the regression by relaxing the stricter error reporting introduced
by commit 369aaa4178e7 and make an exception for the `docker:`
transport.  Note that invalid input would still be caught a couple of
lines below.

Signed-off-by: Valentin Rothberg <rothberg@redhat.com>