Use defer bar.Abort() to ensure every progress bar terminates
mpb.Progress.Wait() blocks until all progress bars are either
marked as completed, aborted, or the context is canceled.

We currently rely on canceling a context, but we want to stop doing that.

To prepare, use defer bar.Abort() to ensure every mpb.Bar terminates.

For now, this should have little to no effect. (Note that bar.Abort()
does nothing after the bar is marked as completed; so we only move the
cancelation forward from canceling the progress pool.) It is a separate
commit to separate the code re-indentation from the other part,
which is less code change but with more impact.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>