Don't compute the (compressed) digest in PutBlob if the caller provides it
We rely on the caller's validation, which is implied but not quite
explicitly guaranteed by the PutBlob documentation; anyway it works
with copy.Image in practice.

The digest.Algorithm().Available() checks go away completely again;
it's the caller's concern how to validate the digest; from now on,
we use the values only for hash map indexing.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>