Modify findAuthentication to use a string key instead of a reference.Named
A reference.Named cannot represent a "docker.io/vendor" namespace
in its _string format_; it can actually be created as a Go value, but
such a value violates our expectations (converting to a string and back
would result in "docker.io/library/vendor").

The "login" implementation needs to read the credentials for an arbitrary
key, possibly including such a "docker.io/vendor" key; we plan to support
that using the text GetCredentials function.

So, build the underlying support for that by modifying the findAuthentication
implementation to work with a string key. For now, GetCredentials
still assumes the input is a registry, and does not work with namespaced keys.

Also rename findAuthentication to findCredentialsInFile, continuing
to eradicate the "Authentication" naming for credentials.

Should not change behavior.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>