Allow using namespaced keys in GetCredentials and GetAuthentication
As described previously, the login code needs to look up values
scoped to docker.io/vendor, which is problematic to represent as
a reference.Named; so, accept namespaced string keys.

Also consolidate/update the API documentation.

The tests have been only minimally updated for consistency, and
will be consolidated later.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>