store.DeleteImage: correctly account for mapped layers
When walking the list of layers in an image to decide which of them to
remove as part of removing that image, we check that each candidate
layer is a) not being used as the top layer for any other images, and b)
that the only layer that lists a candidate layer as its parent layer is
either the top layer of the image we're removing, or the previous layer
that we decided should be removed.

When the image has multiple top layers because some of them are
alternate top layers which use different ID mappings, that second check
didn't account for those mapped layers as being layers that shouldn't
prevent us from removing a candidate layer.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>